Extract CSipHasher to it's own file in crypto/ directory.
This is a move-only commit with the exception of changes to includes.

Cherry-picked from: 4fb789e9b2ffdf48fd50293b3982b3fce4d5fbdf
