+++ src/blockencodings.cpp
@@ -6,6 +6,7 @@
 #include "consensus/consensus.h"
 #include "consensus/validation.h"
 #include "chainparams.h"
+#include "crypto/siphash.h"
 #include "hash.h"
 #include "random.h"
 #include "streams.h"
