+++ src/net.h
@@ -12,6 +12,7 @@
 #include "amount.h"
 #include "bloom.h"
 #include "compat.h"
+#include "crypto/siphash.h"
 #include "hash.h"
 #include "netaddress.h"
 #include "protocol.h"
