+++ src/txmempool.h
@@ -16,6 +16,7 @@
 
 #include "amount.h"
 #include "coins.h"
+#include "crypto/siphash.h"
 #include "indirectmap.h"
 #include "primitives/transaction.h"
 #include "sync.h"
